requirements.txt
opencv_python==4.2.0.34
matplotlib==2.1.1
tensorflow_gpu==1.14.0
scikit_image==0.17.2
numpy==1.13.3
osgeo==0.0.0
Pillow==8.0.1
skimage==0.0
tensorflow==2.4.0
